added ga code
