+++ layouts/event.vue
@@ -364,6 +364,7 @@ import OldSite from '~/components/common/oldsite'
 import Theme from '~/components/common/theme'
 import Upgrade from '~/components/common/upgrade'
 import userUtils from '~/utility/userApps'
+const murmurhash = require('murmurhash')
 export default {
   middleware: ['auth', 'authorization'],
   components: {
@@ -461,6 +462,17 @@ export default {
     this.allowUser = userRoles.length === 1 && userRoles.includes('$orguser')
     this.allowUpgrade = userRoles.includes('$orgowner')
     window.addEventListener('message', this.messageReceived, false)
+    const checkId = murmurhash.v2(
+      this.$auth.user.data.email,
+      this.$config.seedValue
+    )
+    setTimeout(() => {
+      if (window && window.ga) {
+        window.ga('create', this.$config.gaTrackingCode, 'auto')
+        window.ga('set', 'userId', checkId)
+        window.ga('send', 'pageview')
+      }
+    }, 1500)
   },
   beforeDestroy() {
     window.removeEventListener('message', this.messageReceived)
